# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Trabajo",
        "description": "Documentación de la API desarrollada en Laravel para la bolsa de trabajo.",
        "contact": {
            "name": "Soporte",
            "email": "soporte@tuapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor Laravel local"
        }
    ],
    "paths": {
...
        "/api/testimonios": {
            "get": {
                "tags": [
                    "Testimonios"
                ],
                "summary": "Obtener todos los testimonios",
                "operationId": "aa80e27bf3e93553115cac75c3e35076",
                "responses": {
                    "200": {
                        "description": "Lista de testimonios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "testimonials": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Testimonios"
                ],
                "summary": "Crear un nuevo testimonio",
                "operationId": "c1c2dac55fc1f19257d3852102e44d19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Titulo",
                                    "Comentario"
                                ],
                                "properties": {
                                    "Titulo": {
                                        "type": "string",
                                        "example": "Excelente experiencia"
                                    },
                                    "Comentario": {
                                        "type": "string",
                                        "example": "La plataforma me ayudó a encontrar mi primer trabajo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Testimonio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "testimonial": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/testimonios/{id}/toggle": {
            "put": {
                "tags": [
                    "Testimonios"
                ],
                "summary": "Cambiar el estado de un testimonio",
                "operationId": "57066f27188c6d7b7c6dfbe2c1517336",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del testimonio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado cambiado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string"
                                        },
                                        "estado": {
                                            "type": "string",
                                            "enum": [
                                                "Activo",
                                                "Inactivo"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimonio no encontrado"
                    }
                }
            }
        },
        "/api/testimonios/{id}": {
            "get": {
                "tags": [
                    "Testimonios"
                ],
                "summary": "Obtener un testimonio específico",
                "operationId": "2cc1b60ccc616e4aac3daaa52969d18c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del testimonio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Testimonio encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "testimonial": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimonio no encontrado"
                    }
                }
            }
        },
        "/api/testimonios/validate": {
            "post": {
                "tags": [
                    "Testimonios"
                ],
                "summary": "Validar contenido de un testimonio",
                "operationId": "f35ed91866b4b73e4953e9ae48fccb42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Titulo",
                                    "Comentario"
                                ],
                                "properties": {
                                    "Titulo": {
                                        "type": "string",
                                        "example": "Excelente experiencia"
                                    },
                                    "Comentario": {
                                        "type": "string",
                                        "example": "La plataforma me ayudó a encontrar mi primer trabajo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comentario validado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string"
                                        },
                                        "puntajes": {
                                            "properties": {
                                                "toxicidad": {
                                                    "type": "number"
                                                },
                                                "insulto": {
                                                    "type": "number"
                                                },
                                                "profanidad": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Contenido inapropiado"
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            }
        },
...
    },
...
}